Reduced code redundancy
